move result-reason into its own blueprint
