+++ etl-pipeline/api/blueprints/__init__.py
@@ -1,4 +1,5 @@
 from .chat import chat_blueprint
+from .result_reason import result_reason_blueprint
 from .drbCitation import citation
 from .drbCollection import collection, collections
 from .drbEdition import edition, editions
